refactor: Improve manga list handling and data consistency
This commit introduces several improvements to how the application fetches, stores, and displays manga and anime lists, focusing on data consistency and error handling. It also includes minor data type adjustments and database versioning.

**Key Changes:**

- **Manga List Fetching and Caching (`MangaRepository`, `MangaRemoteDataSource`, `MangaLocalDataSource`):**
    - The `MangaRepository` now prioritizes emitting the locally cached manga list first. If the local cache is empty, it then attempts to fetch the list from the remote data source and caches it.
    - Added error handling (try-catch blocks with Timber logging) in `MangaRepository` and `MangaRemoteDataSource` during remote manga list fetching to prevent crashes and log issues.
    - `MangaRemoteDataSource.getMangaList()` now returns an empty list on failure instead of potentially crashing.
    - Added logging in `MangaLocalDataSource` to track the number of entities when `getMangaListFlow` is called.

- **Anime List Caching (`AnimeRepository`, `AnimeLocalDataSource`):**
    - The `AnimeRepository` now emits the local flow immediately and then checks if the local cache is empty to fetch from remote if needed. This ensures the UI can display cached data faster.
    - Added error handling (try-catch block with Timber logging) in `AnimeLocalDataSource.saveAnimeList()` to gracefully handle potential database errors.

- **ViewModel (`MyListViewModel`):**
    - Refactored `observeLocal()` to use `coroutineScope` and launch separate coroutines for observing anime and manga list flows concurrently. This can improve responsiveness.
    - Changed `listSwitchChecked` from `remember { viewModel.listSwitchChecked }` to `viewModel.listSwitchChecked` in `MyListScreen` for potentially better state management.

- **Data Models:**
    - **`UpdateMangaResponse.kt`**: Added a new data class to represent the response when updating manga list status.
    - **`MangaListResponse.kt`**: Changed the type of `mean` from `Double` to `Float?` to allow for null values, which can occur in API responses.
    - **`TokenResponse.kt`**: Changed the type of `expiresIn` from `Int` to `Long` for better precision with token expiration times.

- **Database:**
    - **`AppDatabase.kt`**: Incremented the database `version` from 1 to 2. This is necessary due to schema changes or to trigger migrations if implemented.

+++ app/src/main/java/com/branwen/mal/models/TokenResponse.kt
@@ -3,7 +3,7 @@ package com.branwen.mal.models
 import com.squareup.moshi.Json
 
 data class TokenResponse(
-    @param:Json(name = "expires_in") val expiresIn: Int,
+    @param:Json(name = "expires_in") val expiresIn: Long,
     @param:Json(name = "token_type") val tokenType: String,
     @param:Json(name = "access_token") val accessToken: String,
     @param:Json(name = "refresh_token") val refreshToken: String
